refactor: 動画関連の処理はトピック関連のコンポーネントから外部化
