+++ components/organisms/TopicPreviewDialog.tsx
@@ -1,9 +1,7 @@
-import { useEffect } from "react";
 import Dialog from "@mui/material/Dialog";
 import TopicViewerContent from "$organisms/TopicViewerContent";
 import type { TopicSchema } from "$server/models/topic";
 import useCardStyles from "$styles/card";
-import { useVideoAtom } from "$store/video";
 
 type Props = {
   topic: TopicSchema;
@@ -14,10 +12,6 @@ type Props = {
 export default function TopicPreviewDialog(props: Props) {
   const cardClasses = useCardStyles();
   const { topic, open, onClose } = props;
-  const { video } = useVideoAtom();
-  useEffect(() => {
-    video.clear();
-  }, [video]);
   return (
     <Dialog
       open={open}
