+++ components/organisms/Video/index.tsx
@@ -1,30 +1,92 @@
-import { useMemo } from "react";
-import type { SxProps } from "@mui/system";
+import { useEffect } from "react";
+import usePrevious from "@rooks/use-previous";
+import clsx from "clsx";
+import { css } from "@emotion/css";
 import type { VideoResourceSchema } from "$server/models/videoResource";
-import VideoPlayer from "./VideoPlayer";
-import getVideoInstance from "$utils/video/getVideoInstance";
+import VideoPlayer from "$organisms/Video/VideoPlayer";
+import VideoResource from "$organisms/Video/VideoResource";
+import { NEXT_PUBLIC_VIDEO_MAX_HEIGHT } from "$utils/env";
+import { useVideoAtom } from "$store/video";
+import { useBookAtom } from "$store/book";
+import type { SxProps } from "@mui/system";
 
-type Props = Pick<VideoResourceSchema, "providerUrl" | "url" | "tracks"> & {
-  sx?: SxProps;
+const hidden = css({
+  m: 0,
+  width: 0,
+  "& *": {
+    visibility: "hidden",
+  },
+});
+
+const videoStyle = {
+  "& > *": {
+    /* NOTE: 各動画プレイヤーのレスポンシブ対応により、高さはpaddingTopによってwidthのpercentage分
+     * 確保されるため、heightによる制限ではなくwidthによる制限をおこなう必要がある */
+    // NOTE: 16:9前提になっているが本当はアスペクト比に応じて最大高さを変えたい
+    maxWidth:
+      NEXT_PUBLIC_VIDEO_MAX_HEIGHT === "unset"
+        ? "unset"
+        : `calc(${NEXT_PUBLIC_VIDEO_MAX_HEIGHT} * 16 / 9)`,
+    margin: "0 auto",
+  },
+} as const;
+
+type Props = {
   className?: string;
+  sx?: SxProps;
+  resource: VideoResourceSchema;
   onEnded?: () => void;
-  onDurationChange?: (duration: number) => void;
-  autoplay?: boolean;
 };
 
-export default function Video({
-  providerUrl,
-  url,
-  tracks: resourceTracks,
-  autoplay = false,
-  ...other
-}: Props) {
-  const videoInstance = useMemo(() => {
-    return getVideoInstance(
-      { providerUrl, url, tracks: resourceTracks },
-      autoplay
-    );
-  }, [providerUrl, url, autoplay, resourceTracks]);
-
-  return <VideoPlayer videoInstance={videoInstance} {...other} />;
+export default function Video({ className, sx, resource, onEnded }: Props) {
+  const { video, updateVideo } = useVideoAtom();
+  const { book, itemIndex, itemExists } = useBookAtom();
+  const prevItemIndex = usePrevious(itemIndex);
+  useEffect(() => {
+    if (!book) {
+      video.clear();
+      return;
+    }
+    updateVideo(book.sections);
+  }, [book, video, updateVideo]);
+  useEffect(() => {
+    if (prevItemIndex?.some((v, i) => v !== itemIndex[i])) {
+      video.get(itemExists(prevItemIndex)?.resource.url ?? "")?.player.pause();
+    }
+    const videoInstance = video.get(itemExists(itemIndex)?.resource.url ?? "");
+    if (!videoInstance) return;
+    if (videoInstance.type == "vimeo") {
+      void videoInstance.player.setCurrentTime(0);
+      void videoInstance.player.play();
+    } else {
+      videoInstance.player.ready(() => {
+        videoInstance.player.currentTime(0);
+        void videoInstance.player.play();
+      });
+    }
+  }, [video, itemExists, prevItemIndex, itemIndex]);
+  return (
+    <>
+      {Array.from(video.values()).map((videoInstance) => (
+        <VideoPlayer
+          key={videoInstance.url}
+          className={clsx(className, {
+            [hidden]: resource.url != videoInstance.url,
+          })}
+          sx={{ ...videoStyle, ...sx }}
+          videoInstance={videoInstance}
+          onEnded={onEnded}
+        />
+      ))}
+      {video.size === 0 && (
+        <VideoResource
+          className={className}
+          sx={{ ...videoStyle, ...sx }}
+          {...resource}
+          onEnded={onEnded}
+          autoplay
+        />
+      )}
+    </>
+  );
 }
